Restructured app
- Created app, assets, and components directories
- Removed unused "prop-types" dependency
- Removed unnecessary serviceWorker
