+++ client/package.json
@@ -8,7 +8,6 @@
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "lodash": "^4.17.11",
-    "prop-types": "^15.7.2",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-fontawesome": "^1.6.1",
