+++ client/src/app/projects/ProjectSummaryRenderer.js
