+++ client/src/app/projects/summaries/JubileeProject.js
