+++ client/src/app/skills/Skill.js
