+++ client/src/app/skills/SkillType.css
