+++ client/src/assets/images/maze-game.png
